re-enable auth_external_SUITE
Move the sample_external_auth.py from rel/files to the suite data dir.
At this moment it's used only in small tests and there is no need to
add this file to any release
