+++ rebar.config
@@ -59,8 +59,6 @@
         {extended_start_script, false},
 
         {overlay, [
-                   {copy, "rel/files/sample_external_auth.py", "priv/sample_external_auth.py"},
-
                    {mkdir, "priv/ssl"},
                    {copy, "fake_cert.pem", "priv/ssl/fake_cert.pem"},
                    {copy, "fake_key.pem", "priv/ssl/fake_key.pem"},
